~move common multiplatform configs to the convention plugin
