+++ bytestring/build.gradle.kts
@@ -1,5 +1,5 @@
+import Multiplatform_lib_conventions_gradle.IoMultiplatformExtension
 import org.jetbrains.dokka.gradle.DokkaTaskPartial
-import org.jetbrains.kotlin.gradle.plugin.KotlinSourceSet
 
 plugins {
     id("multiplatform-lib-conventions")
@@ -8,16 +8,6 @@ plugins {
 }
 
 kotlin {
-    jvm {
-        jvmToolchain {
-            languageVersion.set(JavaLanguageVersion.of(libs.versions.java.get()))
-        }
-        withJava()
-        testRuns["test"].executionTask.configure {
-            useJUnitPlatform()
-        }
-    }
-
     js(IR) {
         nodejs {
             testTask {
@@ -28,42 +18,15 @@ kotlin {
         }
         browser {
             testTask {
-                filter.setExcludePatterns("*SmokeFileTest*")
                 useMocha {
                     timeout = "30s"
                 }
             }
         }
     }
-
-    sourceSets {
-        val commonMain by getting
-        val commonTest by getting {
-            dependencies {
-                implementation(kotlin("test"))
-            }
-        }
-    }
-
-    explicitApi()
-    sourceSets.configureEach {
-        configureSourceSet()
-    }
 }
 
-fun KotlinSourceSet.configureSourceSet() {
-    val srcDir = if (name.endsWith("Main")) "src" else "test"
-    val platform = name.dropLast(4)
-    kotlin.srcDir("$platform/$srcDir")
-    if (name == "jvmMain") {
-        resources.srcDir("$platform/resources")
-    } else if (name == "jvmTest") {
-        resources.srcDir("$platform/test-resources")
-    }
-    languageSettings {
-        progressiveMode = true
-    }
-}
+
 
 tasks.withType<DokkaTaskPartial>().configureEach {
     dokkaSourceSets.configureEach {
@@ -76,4 +39,8 @@ tasks.withType<DokkaTaskPartial>().configureEach {
 
         samples.from("common/test/samples/samples.kt")
     }
+}
+
+extensions.configure<IoMultiplatformExtension> {
+    javaVersion.set(JavaLanguageVersion.of(libs.versions.java.get()))
 }
